+++ conduit/plugins/exporters/postgresql/postgresql_exporter_test.go
@@ -15,8 +15,8 @@ import (
 	"github.com/algorand/indexer/conduit/plugins"
 	"github.com/algorand/indexer/conduit/plugins/exporters"
 	"github.com/algorand/indexer/conduit/plugins/exporters/postgresql/util"
+	"github.com/algorand/indexer/conduit/plugins/tools/testutil"
 	_ "github.com/algorand/indexer/idb/dummy"
-	testutil "github.com/algorand/indexer/util/test"
 )
 
 var pgsqlConstructor = exporters.ExporterConstructorFunc(func() exporters.Exporter {
