+++ conduit/plugins/importers/algod/mock_algod_test.go
@@ -1,4 +1,4 @@
-package test
+package algodimporter
 
 import (
 	"net/http"
